Loss Protected Fixes

+++ CCP/android/reference_wallet/fermat-ccp-android-loss-protected-bitcoin-wallet-bitdubai/src/main/java/com/bitdubai/reference_niche_wallet/loss_protected_wallet/common/adapters/PaymentRequestHistoryAdapter.java
@@ -209,7 +209,7 @@ public class PaymentRequestHistoryAdapter  extends FermatAdapter<LossProtectedPa
                     try {
 
                         //verify loss protected settings
-                        if(appSession.getData(SessionConstant.ACTUAL_EXCHANGE_RATE) != 0){
+                        if(((double)appSession.getData(SessionConstant.ACTUAL_EXCHANGE_RATE)) != 0){
